requirements.txt
timm>=0.5.4
torch>=1.10.1
torchvision>=0.11.2
torchmetrics>=0.7.0
numpy>=1.22.1
Pillow>=9.0.0
pandas>=1.3.5
pytorch-lightning>=1.5.8
tensorboard>=2.7.0
matplotlib>=3.5.1